Log when we don't get a valid response.
There are cases where you might not get a valid response, due to network
issues, SSL problems, et cetera. This captures those cases, leaning on
http-kit's :error key if there isn't a valid :body in the response. This
then gets logged.

As is, the code will register a null pointer, as there won't be a body
to deserialize, and `json/read-str` will fail.